(DIO-944) Merge the Scraper and Web Front End
Modify the scraper to run at intervals in a thread and merge the Web
front end into the same program.

Add Last Updated to the Products data returned to the react app, as this
is no longer statically built in the Jenkins build.

Remove unused functions and code.

Merge in changes from #15 and #17 as suggested by dependabot.