(DIO-944) Restructure into backend/frontend
Put the main.go and its associated lib and configuration files into a
backend folder so we can have separate Dockerfile's for the frontend
and backend.

Re-work Dockerfile to get the backend app working (finally)

go.sum/go.mod still need to be in the root folder, with the docker build
command run in that context.

Add .dockerignore to filter out crud extending the docker build time.

Remove the Links Functionality altogether as it wasn't being used in the
prod .toml file and think it can be better implemented anyway in the
sidebar.

The configuration has been simplified to make it more consistent. viper
and pflag have been removed for the moment, but viper may well be
introducted later on as its a good model for configuration.